# Conflicts:
#	server/src/main/java/com/ifmo/adda/controller/OrderController.kt
#	server/src/main/java/com/ifmo/adda/dao/Cart.kt
#	server/src/main/java/com/ifmo/adda/dto/CartDto.kt
#	server/src/main/java/com/ifmo/adda/dto/OrderDto.kt
#	server/src/main/java/com/ifmo/adda/repository/CartRepository.kt

+++ server/src/main/java/com/ifmo/adda/dao/CustomOrder.kt
@@ -13,7 +13,7 @@ data class CustomOrder(
         var id: Int? = null,
 
         @Column(nullable = false)
-        val client: Int,
+        val client: BigInteger,
 
         @Column(nullable = true)
         val description: String,
