+++ server/src/main/java/com/ifmo/adda/dao/Order.kt
@@ -1,5 +1,6 @@
 package com.ifmo.adda.dao
 
+import java.math.BigInteger
 import java.time.Instant
 import javax.persistence.*
 
@@ -12,7 +13,7 @@ data class Order(
     var id: Int? = null,
 
     @Column(nullable = false)
-    val client: Int,
+    val client: BigInteger,
 
     @Column(nullable = false)
     var dateOfOrder: Instant,
