+++ server/src/main/java/com/ifmo/adda/repository/CustomOrdersRepository.kt
@@ -4,12 +4,13 @@ import com.ifmo.adda.dao.CustomOrder
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
+import java.math.BigInteger
 
-interface CustomOrdersRepository : JpaRepository<CustomOrder, Int> {
+interface CustomOrdersRepository : JpaRepository<CustomOrder, BigInteger> {
 
     @Query(
             "SELECT DISTINCT custom_order_item FROM CustomOrder custom_order_item WHERE custom_order_item.client = :clientId"
     )
-    fun findCustomForClient(@Param("clientId") clientId: Int): List<CustomOrder>
+    fun findCustomForClient(@Param("clientId") clientId: BigInteger): List<CustomOrder>
 
 }
