+++ server/src/main/java/com/ifmo/adda/repository/OrdersRepository.kt
@@ -6,15 +6,16 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
+import java.math.BigInteger
 
-interface OrdersRepository : JpaRepository<Order, Int> {
+interface OrdersRepository : JpaRepository<Order, BigInteger> {
 
     @Query(
             "SELECT DISTINCT order_item FROM Order order_item " +
                     "INNER JOIN FETCH order_item.products as products " +
                     "WHERE order_item.client = :clientId"
     )
-    fun findNormalForClient(@Param("clientId") clientId: Int): List<Order>
+    fun findNormalForClient(@Param("clientId") clientId: BigInteger): List<Order>
 
     @Modifying
     @Query(
